Basic History Package Refactoring
